requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.3
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-extensions==3.1.5
django-filter==21.1
django-timezone-field==5.0
django4-background-tasks==1.2.7
djangorestframework==3.13.1
importlib-metadata==4.11.3
kombu==5.2.4
Markdown==3.3.6
packaging==21.3
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
six==1.16.0
smpplib==2.2.1
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.7.0
